First Commit
An object detection test file using built in haar cascades, another object detection file meant for use with final robot (No can and bottle cascades are included, will include those once I get them finished in the next day or so here), and basic layout for other python source files, with some ideas for how the robot's overall process can be run. I also included a mockup for the basic control panel webpage that I made early on in case it could be of use in designing a final one.